+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - New: option to set an alternative base URL for the API calls
 - New: run diagnostics (network, accounts, settings, gem)
+- New: bypass authorization webpage by providing the token inline
 - Fixed: situations when no user authorized
 - Fixed: --unauthorize if no valid username
 - Deprecated: Ayadn 1.x migration tools